BlogLite v 2.0/requirements.txt
flask
Flask-SQLAlchemy
flask-security-too
flask-restful
flask_bcrypt
Flask-Caching
flask_cors
redis
celery[redis]
email_validator
Flask-Mail
weasyprint
matplotlib
numpy